Finished Team Manager
Finished Team Manager React App that shows a list of all players and their positions on the main page. There is then a page to add a new player to the database with validations. There is a delete button with a popup for confirmation and a player status page with buttons to change the player's playing status for each of three games.